requirements.txt
aiofiles==24.1.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==2.1.0
attrs==25.3.0
Automat==25.4.16
backports.asyncio.runner==1.2.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.12.0
itemloaders==1.3.2
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
Levenshtein==0.27.1
lxml==6.0.0
lxml_html_clean==0.4.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql==0.0.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
orjson==3.11.0
packaging==25.0
pandas==2.3.1
parsel==1.10.0
phonenumbers==9.0.10
pillow==11.3.0
playwright==1.54.0
pluggy==1.6.0
Protego==0.5.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
Scrapy==2.13.3
sentry-sdk==2.33.2
service-identity==24.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
tinysegmenter==0.3
tldextract==5.3.0
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
w3lib==2.3.1
watchfiles==1.1.0
websockets==15.0.1
zope.interface==7.2
